Resolve #147

+++ src/main/java/com/model/managers/CourseManager.java
@@ -8,14 +8,14 @@ import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
 import com.model.Course;
+import com.model.DataHandlers.DataLoader;
+import com.model.DataHandlers.DataWriter;
 import com.model.InstrumentType;
 import com.model.Lesson;
 import com.model.OperationResult;
 import com.model.SavableList;
 import com.model.Student;
 import com.model.Teacher;
-import com.model.DataHandlers.DataLoader;
-import com.model.DataHandlers.DataWriter;
 
 /**
  * A manager for all courses
@@ -163,7 +163,7 @@ public class CourseManager implements SavableList<Course> {
 			if (numTimesObj==null) { //safe against nullptr;
 				numTimesObj = 1;
 			}
-			int numberOfTimes = (Integer)numTimesObj;
+			int numberOfTimes = ((Long)numTimesObj).intValue();
 
 			Lesson l = new Lesson(lessonID, lessonTitle, songID,InstrumentType.valueOf(instrument.toUpperCase()),numberOfTimes);
 			lessons.add(l);
